Add some kdocs
