+++ src/main/kotlin/io/libp2p/core/multiformats/Multiaddr.kt
@@ -6,16 +6,45 @@ import io.libp2p.etc.types.toByteBuf
 import io.netty.buffer.ByteBuf
 import io.netty.buffer.Unpooled
 
+/**
+ * Class implements Multiaddress concept: https://github.com/multiformats/multiaddr
+ *
+ * Multiaddress is basically the chain of components like `protocol: value` pairs
+ * (value is optional)
+ *
+ * It's string representation is `/protocol/value/protocol/value/...`
+ * E.g. `/ip4/127.0.0.1/tcp/1234` which means TCP socket on port `1234` on local host
+ *
+ * @param components: generic Multiaddress representation which is a chain of 'components'
+ * represented as a known [Protocol] and its value (if any) serialized to bytes according
+ * to this protocol rule
+ */
 class Multiaddr(val components: List<Pair<Protocol, ByteArray>>) {
 
+    /**
+     * Creates instance from the string representation
+     */
     constructor(addr: String) : this(parseString(addr))
 
+    /**
+     * Creates instance from serialized form from [ByteBuf]
+     */
     constructor(bytes: ByteBuf) : this(parseBytes(bytes))
+    /**
+     * Creates instance from serialized form from [ByteBuf]
+     */
     constructor(bytes: ByteArray) : this(parseBytes(bytes.toByteBuf()))
 
+    /**
+     * Returns [components] in a human readable form where each protocol value
+     * is deserialized and represented as String
+     */
     fun getStringComponents(): List<Pair<Protocol, String?>> =
         components.map { p -> p.first to if (p.first.size == 0) null else p.first.bytesToAddress(p.second) }
 
+    /**
+     * Serializes this instance to supplied [ByteBuf]
+     */
     fun writeBytes(buf: ByteBuf): ByteBuf {
         for (component in components) {
             buf.writeBytes(component.first.encoded)
@@ -24,8 +53,16 @@ class Multiaddr(val components: List<Pair<Protocol, ByteArray>>) {
         return buf
     }
 
+    /**
+     * Returns serialized form as [ByteArray]
+     */
     fun getBytes(): ByteArray = writeBytes(Unpooled.buffer()).toByteArray()
 
+    /**
+     * Returns the string representation of this multiaddress
+     * Note that `Multiaddress(strAddr).toString` is not always equal to `strAddr`
+     * (e.g. `/ip6/::1` can be converted to `/ip6/0:0:0:0:0:0:0:1`)
+     */
     override fun toString(): String = getStringComponents().joinToString(separator = "") { p ->
         "/" + p.first.typeName + if (p.second != null) "/" + p.second else "" }
