+++ src/main/kotlin/io/libp2p/etc/Attributes.kt
@@ -1,5 +1,7 @@
-package io.libp2p.core
+package io.libp2p.etc
 
+import io.libp2p.core.Connection
+import io.libp2p.core.Stream
 import io.libp2p.core.mux.StreamMuxer
 import io.libp2p.core.security.SecureChannel
 import io.netty.channel.Channel
