+++ src/test/kotlin/io/libp2p/tools/TestChannel.kt
@@ -1,9 +1,9 @@
 package io.libp2p.tools
 
 import com.google.common.util.concurrent.ThreadFactoryBuilder
-import io.libp2p.core.CONNECTION
 import io.libp2p.core.Connection
-import io.libp2p.core.IS_INITIATOR
+import io.libp2p.etc.CONNECTION
+import io.libp2p.etc.IS_INITIATOR
 import io.libp2p.etc.types.lazyVar
 import io.libp2p.etc.util.netty.nettyInitializer
 import io.netty.channel.ChannelHandler
